Доработать программу согласно постман тестам

+++ main-service/src/main/java/ru/practicum/ewm/event/dto/EventAdminFilter.java
@@ -14,9 +14,9 @@ import java.util.List;
 @Builder
 @FieldDefaults(level = AccessLevel.PRIVATE)
 public class EventAdminFilter {
-    List<Integer> users;
+    List<Long> users;
     List<State> states;
-    List<Integer> categories;
+    List<Long> categories;
     LocalDateTime rangeStart;
     LocalDateTime rangeEnd;
     Boolean paid;
